tools/headlessxi-2/requirements.txt
blowfish==0.6.1
Brotli==1.0.9
click==8.1.3
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.12
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
kazoo==2.8.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
plotly==5.10.0
six==1.16.0
tenacity==8.0.1
Werkzeug==2.2.2
zipp==3.8.1
